refactor: migrate dashboard components to app directory and implement new SessionTime, Injuries, and Availability components
