+++ src/app/dashboard/components/EquipmentAccess.tsx
