+++ src/app/dashboard/components/Injuries.tsx
