+++ src/app/dashboard/components/UserAnswers.tsx
@@ -2,7 +2,7 @@
 
 import { useUserAnswers } from "@/hooks/useUserAnswers";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
-import { AnswerCard } from "@/components/dashboard/AnswerCard";
+import { AnswerCard } from "@/app/dashboard/components/AnswerCard";
 import type { UserAnswer } from "@/types/UserAnswer"; // Importar el tipo centralizado
 
 /**
